+++ Win32Emu/Win32/Modules/Gdi32Module.cs
@@ -367,7 +367,7 @@ namespace Win32Emu.Win32.Modules
 		private uint GetStockObject(int stockObjectId)
 		{
 			// Validate stock object ID
-			if (stockObjectId is < NativeTypes.StockObject.WHITE_BRUSH or > NativeTypes.StockObject.DC_PEN)
+			if (stockObjectId is < (int)NativeTypes.StockObject.WHITE_BRUSH or > (int)NativeTypes.StockObject.DC_PEN)
 			{
 				_logger.LogInformation("[Gdi32] GetStockObject: Invalid stock object ID {StockObjectId}", stockObjectId);
 				return 0;
